Don't process new_code if we have filled output buffer.
When decoding an LZW stream, after processing a burst, we attempt to
process the next code, but it is possible that out is already filled and
the remaining data in the buffer does not consist of valid LZW codes.

This commit allows me to decode the image I have, but I don't think it
is correctly setting code_link (I think my use case only calls advance
once per chunk, so code_link is never used).

I tested it on an image that previously crashed, and no longer crashes.
To check it parsed correctly, I re-saved the parsed image, and used
imagemagick's compare to confirm that the resulting image is the same as
the source.